Create DROP INDEX statements during conversion
We can no longer count on GitLab to drop indexes for us in the restore script.
Because I cannot be bothered right now to rewrite db_converter.py to drop
support for writing to stdout, and allow seeking, and inserting lines at the
beginning, I added a small Ed script that moves the DROP INDEX statements to a
better spot than the end of the DB import.

I hope Ed does not choke on multi-GB files.